additional review fixes
