+++ test/e2e/page-objects/pages/home/activity-tab.ts
@@ -123,12 +123,23 @@ class ActivityTab extends HomePage {
    * This function checks if the specified number of confirmed transactions are displayed in the activity list on homepage.
    * It waits up to 10 seconds for the expected number of confirmed transactions to be visible.
    *
-   * @param expectedNumber - The number of confirmed transactions expected to be displayed in activity list. Defaults to 1.
+   * @param expectedNumber - The number of confirmed transactions expected to be displayed in activity list. Defaults to 1. Pass 0 to assert none are shown.
    * @returns A promise that resolves if the expected number of confirmed transactions is displayed within the timeout period.
    */
   async checkConfirmedTxNumberDisplayedInActivity(
     expectedNumber: number = 1,
   ): Promise<void> {
+    if (expectedNumber === 0) {
+      console.log(
+        'Verify no confirmed transactions are displayed in activity list',
+      );
+      await this.driver.assertElementNotPresent(this.confirmedTransactions);
+      console.log(
+        'No confirmed transactions found in activity list on homepage',
+      );
+      return;
+    }
+
     console.log(
       `Wait for ${expectedNumber} confirmed transactions to be displayed in activity list`,
     );
